backend-microservice/hybrid-model-microservice/requirements.txt
flask
numpy
prometheus_client 
pillow
gunicorn
